Upgraded form system work in progress
Added form::open and form::close methods. Created a system to attach a model to a form and have it injected with data into an action when submitting. Began to refactor and stablize form code.

Todo: Commenting needs to be improved and the rest of the form methods need to be upgraded to work with the new system. Continued refactoring.