Refactored form::checkbox method
Cleaned up form::checkbox method to work better and be cleaner. Removed the option to turn off the zerofill functionality. Zerofill functionality adds a post parameter of false for checkboxes that are unchecked using a hidden input field. None of my projects have ever needed to turn it off so I'm removing the option.