url::context() method
This method returns the name of the current rendering context so that scripts can key off of it in controllers and view logic. Sometimes it may be advantageous to save a couple database requests if the full view isn't being rendered or to move some markup around if the context is within a popup or other dynamic placement. url::context is just a forward to url::request but I would prefer to keep the api independant of specific url parameters for future implementation flexibility.